requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.7.0.68
pandas==1.1.5
Pillow==8.4.0
pyntcloud==0.1.5
pythreejs==2.3.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
tensorflow==1.13.2
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.64.1

